use the same confluence loader with different connection logic
